backend/requirements.txt
alembic==1.14.1
edgartools==5.28.1
fastapi==0.115.12
itsdangerous==2.2.0
numpy>=1.26.0
pandas>=2.2.0
psycopg==3.3.3
psycopg-binary==3.3.3
pydantic==2.11.3
pydantic-settings==2.8.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
sqlalchemy==2.0.40
uvicorn[standard]==0.34.1
yfinance==0.2.55
